+++ packages/data-model/src/entity/AcDbTable.ts
@@ -415,6 +415,7 @@ export class AcDbTable extends AcDbBlockReference {
     quaternion.setFromAxisAngle(AcGeVector3d.Z_AXIS, this.rotation)
     _tmpMatrix.compose(this.position, quaternion, this.scaleFactors)
     group.applyMatrix(_tmpMatrix)
+    this.attachEntityInfo(group)
     return group
   }