+++ packages/data-model/src/entity/AcDbViewport.ts
@@ -211,6 +211,7 @@ export class AcDbViewport extends AcDbEntity {
     ) {
       const viewport = this.toGiViewport()
       const group = renderer.group(this.createViewportRect(viewport, renderer))
+      this.attachEntityInfo(group)
       return group
     }
     return undefined